Fix serve_multiple's stop_event
Just a simple .wait() on the stop_event is enough to make it wait until
SIGINT.

The problem with the test is that after_start is set _before_ the signal
handler in the child process, and after_start calls stop_event.set(),
before the signal is registered. Simply reordering the lines fixes that.